clean up models; reduce down to 2; rename
